Move old skypy.core module to new skypy.pipeline.driver module
