+++ skypy/pipeline/__init__.py
@@ -1,0 +1,4 @@
+"""
+This module provides methods to pipeline together multiple models with
+dependencies and handle their outputs.
+"""
